Updates12

+++ src/models.jl
@@ -1,6 +1,7 @@
 using Lux
 using Printf
 using StableRNGs
+using Random
 using CRC32c
 
 scaled_tanh(x::T) where T = tanh(x)/2 .+ T(0.5)
@@ -58,14 +59,20 @@ function compute_loss(ŷ, y, w)
 end
 
 #this is just random; replace with something meaningful
-accuracy(y_pred, y_true) = mean(sqrt.(sum(abs2,y_pred[:,end-10:end,:] .- y_true[:,end-10:end,:],dims=1)))
+function accuracy(y_pred, y_true) 
+    Δ = sqrt(sum(abs2,y_pred[:,end-10:end,:] .- y_true[:,end-10:end,:]))
+    Δp = sqrt(sum(abs2,y_pred[:,end-10:end,:])) .+ sqrt.(sum(abs2,y_true[:,end-10:end,:]))
+    mean(Δ/Δp)
+end
 
 function train_model(model, x::AbstractArray{Float32,3},y::AbstractArray{Float32,3},z::AbstractArray{Float32,3};kwargs...)
     train_model(model, ()->(x,y,z), accuracy;kwargs...)
 end
 
-function train_model(model, data_provider, accuracy_func::Function=accuracy, perf_func=accuracy_func;nepochs=25, accuracy_threshold=0.9f0,save_file="model_state.jld2",redo=false, learning_rate=0.01f0, freeze_input=false, rseed=12345, h=zero(UInt64), load_only=false)
-    rng=StableRNG(rseed)
+function train_model(model, data_provider, accuracy_func::Function=accuracy, perf_func=accuracy_func;model_params=nothing, model_state=nothing, nepochs=25, accuracy_threshold=0.9f0,save_file="model_state.jld2",redo=false, learning_rate=0.01f0, freeze_input=false, rseed=12345, h=zero(UInt32), load_only=false)
+    #rng=StableRNG(rseed)
+    rng = Random.default_rng()
+    Random.seed!(rseed)
     # this is hacking; there should be a general way of getting this
     nhidden = model.layers.rnn_cell.out_dims
     output_nonlinearity = model.layers.classifier.activation
@@ -104,9 +111,24 @@ function train_model(model, data_provider, accuracy_func::Function=accuracy, per
     hs = string(h, base=16)
     fname = replace(save_file, ".jld2"=> "_$(hs).jld2")
     logfile = replace(save_file, ".jld2"=> "_log_$(hs).csv")
-    if isfile(fname) && !redo
+    # check if logfile already exists and check progress
+    n_epochs_remaining = nepochs
+    if isfile(logfile) && !redo
+        n_epochs_remaining = open(logfile,"r") do fid
+            lines = readlines(fid)
+            _nepochs = 0
+            if length(lines) > 1
+                _nepochs = parse(Int64, first(split(lines[end],',')))
+            end
+            nepochs - _nepochs
+        end
+    end
+    if (model_params !== nothing) && (model_state !== nothing)
+        _ps = model_params
+        _st = model_state
+    elseif isfile(fname) && !redo
         _ps,_st = JLD2.load(fname, "params","state")
-        if load_only
+        if (load_only) || (n_epochs_remaining <= 0)
             return _ps, _st
         end
         print(stdin, "File $(fname) already exists. Starting training from previous parameters. To restart from a random state, call with `redo=true`\n")
@@ -131,11 +153,14 @@ function train_model(model, data_provider, accuracy_func::Function=accuracy, per
     prog = Progress(nepochs, "Training model...")
     total_loss0 = 0.0f0
     total_loss_p = typemax(Float32)
-    open(logfile, "w") do _logfile
-        write(_logfile, "epoch,loss,validation_loss,validation_accuracy,validation_performance,total_loss_change")
+    if !isfile(logfile)
+        open(logfile, "w") do _logfile
+            write(_logfile, "epoch,loss,validation_loss,validation_accuracy,validation_performance,total_loss_change")
+        end
     end
-    open(logfile,"a") do _logfile
-        for epoch in 1:nepochs
+    _logfile = open(logfile, "a")
+    try
+        for (ii,epoch) in enumerate((nepochs-n_epochs_remaining+1):nepochs)
             (xt,yt,wt)  = dev.(data_provider())
 
             (_, loss, _, train_state) = Training.single_train_step!(
@@ -144,7 +169,7 @@ function train_model(model, data_provider, accuracy_func::Function=accuracy, per
             total_loss = loss * length(yt)
             total_samples = length(yt)
 
-            if epoch == 1
+            if ii == 1
                 total_loss0 = total_loss
             end
             loss = @sprintf "%4.5f" total_loss / total_samples
@@ -195,8 +220,12 @@ function train_model(model, data_provider, accuracy_func::Function=accuracy, per
                # GC.gc()
             end
         end
+    catch ee
+        rethrow(ee)
+    finally
+        close(_logfile)
+        release_training(fname)
     end
-    release_training(fname)
     return cdev((train_state.parameters, train_state.states))
 end
